[ext-225] rename mappers for repo consistency
